+++ yoseobara/src/main/java/com/final2/yoseobara/controller/request/PostRequestDto.java
@@ -1,13 +1,1 @@
-package com.final2.yoseobara.controller.request;
-
-import lombok.Getter;
-
-@Getter
-public class PostRequestDto {
-
-    private String title;
-    private String nickname;
-    private String content;
-    private Float address;
-
-}
+package com.final2.yoseobara.controller.request;import lombok.Getter;@Getterpublic class PostRequestDto {    private String title;    private String nickname;    private String content;    private Float address;}
