--- yoseobara/src/main/java/com/final2/yoseobara/model/Timestamped.java
@@ -1,24 +1,0 @@
-package com.final2.yoseobara.model;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.EntityListeners;
-import java.time.LocalDateTime;
-
-@Getter // get 함수 자동 생성
-@EntityListeners(AuditingEntityListener.class)// 변경되었을 때 자동으로 기록합니다.
-public abstract class Timestamped {
-
-    @CreatedDate // 최초 생성 시점
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
-    private LocalDateTime createdAt;
-
-    @LastModifiedDate // 마지막 변경 시점
-    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "Asia/Seoul")
-    private LocalDateTime modifiedAt;
-
-}
